+++ ydb/core/testlib/grpc_request/ya.make
@@ -1,0 +1,10 @@
+LIBRARY()
+
+PEERDIR(
+    ydb/core/testlib
+    ydb/core/grpc_services/base
+    ydb/library/ydb_issue
+    ydb/public/api/protos
+)
+
+END()
